plot_agent/langraph_plot_agent/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
clickhouse-driver
pandas
matplotlib
seaborn
streamlit==1.31.1
pillow
openai>=1.0.0
clickhouse_connect>=0.6.8
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.0
langchain-community>=0.0.13
langchain-experimental>=0.0.47
langgraph>=0.0.25
graphviz>=0.20.1